found out you could restore before... so now it works

+++ admin/admin_restore_post.py
@@ -1,4 +1,4 @@
-from flask import render_template, request
+from flask import render_template, request, session, redirect, url_for
 from icecream import ic
 ic.configureOutput(prefix=f'----- | ', includeContext=True)
 
@@ -10,21 +10,35 @@ from app import app
 @app.put("/admin_restore_post")
 def api_admin_restore_post():
     try:
-        post_pk = request.args.get("post", "")
+        user = session.get("user")
+        if not user: return redirect(url_for("login"))
 
-        db, cursor = x.db()
-        q = "UPDATE posts SET post_deleted_at = %s WHERE post_pk = %s" 
-        cursor.execute(q, (0, post_pk))
-        db.commit()
+        if "admin" in user['user_role']:
+            post_pk = request.args.get("post", "")
 
-        new_button = f""" 
-            <a href="/admin_delete_post?post={ post_pk }" mix-delete class="delete" id="admin_post_delete{ post_pk }"> { x.lans("remove_post") } </a>
-        """
-        succes_template = render_template(("global/succes_message.html"), message=x.lans("succes"))
-        return f"""
-        <browser mix-replace='#admin_post_delete{post_pk}'> {new_button}</browser>
-        <browser mix-bottom='#succes_message'>{succes_template}</browser>
-        """
+            db, cursor = x.db()
+            q = "SELECT user_language, user_email, user_first_name, user_last_name, post_message FROM users JOIN posts ON user_pk = user_fk WHERE post_pk = %s"
+            cursor.execute(q, (post_pk, ))
+            users_post = cursor.fetchone()
+            ic(users_post)
+
+            q = "UPDATE posts SET post_deleted_at = %s WHERE post_pk = %s" 
+            cursor.execute(q, (0, post_pk))
+            db.commit()
+
+            x.default_language = users_post["user_language"]
+            message_to_user = render_template("___email_post_restore.html", post=users_post)
+            x.send_email(users_post["user_email"], x.lans("one_of_your_post_has_been_restore"), message_to_user)
+            x.default_language = users_post["user_language"]
+
+            new_button = f""" 
+                <a href="/admin_delete_post?post={ post_pk }" mix-delete class="delete" id="admin_post_delete{ post_pk }"> { x.lans("remove_post") } </a>
+            """
+            succes_template = render_template(("global/succes_message.html"), message=x.lans("succes"))
+            return f"""
+            <browser mix-replace='#admin_post_delete{post_pk}'> {new_button}</browser>
+            <browser mix-bottom='#succes_message'>{succes_template}</browser>
+            """
     except Exception as ex:
         ic(ex)
         error_template = render_template(("global/error_message.html"), message=x.lans("error_please_try_again"))
